+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Edge cases with input kickoff times
+- Remove gametime from screen when deactivating extension
 
 ## 0.0.1 - 2023-07-27